+++ CHANGELOG.md
@@ -2183,3 +2183,4 @@
 - [2026-06-22 09:46] feat(storage): add bi-directional event bus handler for storage
 - [2026-06-22 09:47] test(theme): add unit tests for theme business logic
 - [2026-06-22 09:48] perf(perf): apply RepaintBoundary on animated perf widgets
+- [2026-06-22 10:06] test(navigation): add unit tests for navigation business logic